move metadata to store-api
