+++ src/mito2/src/test_util.rs
@@ -27,14 +27,13 @@ use log_store::raft_engine::log_store::RaftEngineLogStore;
 use log_store::test_util::log_store_util;
 use object_store::services::Fs;
 use object_store::ObjectStore;
-use store_api::metadata::ColumnMetadata;
+use store_api::metadata::{ColumnMetadata, RegionMetadataRef};
 use store_api::storage::RegionId;
 
 use crate::config::MitoConfig;
 use crate::engine::MitoEngine;
 use crate::error::Result;
 use crate::manifest::manager::{RegionManifestManager, RegionManifestOptions};
-use crate::metadata::RegionMetadataRef;
 use crate::request::{CreateRequest, RegionOptions};
 use crate::worker::WorkerGroup;
