+++ src/store-api/Cargo.toml
@@ -17,6 +17,7 @@ datatypes = { workspace = true }
 derive_builder.workspace = true
 futures.workspace = true
 serde.workspace = true
+serde_json = "1.0"
 snafu.workspace = true
 
 [dev-dependencies]
